requirements.txt
alembic==1.9.4
astroid==2.14.2
autopep8==2.0.1
bcrypt==4.0.1
click==8.1.3
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
numpy==1.24.2
pandas==1.5.3
platformdirs==3.0.0
plotly==5.13.1
pycodestyle==2.10.0
pylint==2.16.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.4
tenacity==8.2.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1
WTForm==1.0
WTForms==3.0.1
